requirements.txt
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.16
django-3-jet==1.0.8
django-admin-charts==1.0.0
django-admin-volt==0.0.5
django-bower==5.2.0
django-cors-headers==3.11.0
django-import-export==2.8.0
django-jazzmin==2.4.9
django-money==2.1.1
django-multiselectfield==0.1.12
django-nvd3==0.9.7
django-tinymce==3.4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
djcacheutils==3.0.0
drf-yasg==1.20.0
flake8==4.0.1
Pillow==9.0.1
psycopg2-binary==2.9.3
py-moneyed==1.2
PyJWT==2.4.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-decouple==3.6
python-memcached==1.59
python-nvd3==0.14.2
python-slugify==1.1.4
pytz==2021.3
PyYAML==6.0
requests==2.27.1
sentry-sdk==1.8.0
simplejwt==2.0.1
six==1.16.0
sqlparse
stripe
whitenoise
gunicorn
dj_database_url
google-auth